real-time-examples-main/requirements.txt
boto3
python-dotenv
corelink-client


